+++ src/App.js
@@ -1,28 +1,34 @@
 import React, { useState, useEffect, useCallback } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { useCookies } from 'react-cookie';
 
 import Login from './components/login/Login';
 import Comander from './components/comander/Comander';
 import Header from './components/header/Header';
+import Sidebar from './components/sidebar/Sidebar';
 
 import { doLogout } from './service/neo.service';
-import { themes, manageAutoTheme, ThemeContext } from './global/utils/hooks/theme';
+import { getDBSchema } from './service/schema.service';
 import { cls } from './global/utils';
 
 import './App.css';
+import actions from './global/utils/store/actions';
+import { useAsyncDispatch } from './global/utils/hooks/dispatch';
 
 function App() {
     const [cookies, setCookie] = useCookies(["neo4jDash.sess"]);
-    const [sessionId, setSessionId] = useState(null);
-    const [user, setUser] = useState('');
     const [loading, setLoading] = useState(true);
-    const [theme, setTheme] = useState(() => manageAutoTheme('auto'));
+    const [menu, setMenu] = useState(false);
+    const [theme, user] = useSelector(state => [state.theme, state.user]);
 
-    const loginHandler = useCallback((response) => {
-        setCookie("neo4jDash.sess", JSON.stringify(response));
-        setSessionId(response.sessionId);
-        setUser(response.user);
-    }, [setCookie]);
+    const dispatch = useDispatch();
+    const asyncDispatch = useAsyncDispatch();
+
+    const loginHandler = useCallback((userData) => {
+        setCookie("neo4jDash.sess", JSON.stringify(userData));
+        dispatch(actions.user.setUser(userData));
+        asyncDispatch(actions.db.setProperties, getDBSchema(userData.sessionId));
+    }, [asyncDispatch, dispatch, setCookie]);
 
     useEffect(() => {
         if (cookies["neo4jDash.sess"] && cookies["neo4jDash.sess"].sessionId && loading) {
@@ -31,42 +37,43 @@ function App() {
         setLoading(false);
     }, [cookies, loading, loginHandler, theme]);
 
-    const logoutHandler = () => {
-        logout();
-    }
-
-    const logout = async () => {
-        const logOutResult = await doLogout(sessionId);
-        if (logOutResult) {
-            setSessionId(null);
-            setCookie('neo4jDash.sess', null);
+    useEffect(() => {
+        if (!loading && !user.loggedIn) {
+            async function logout() {
+                await doLogout(user.sessionId);
+            }
+            logout();
+            setCookie("neo4jDash.sess", null);
         }
+    }, [loading, setCookie, user]);
+
+    const toggleMenu = (state) => {
+        setMenu(state);
     }
 
     const render = () => {
         if (loading) {
             return <em className={cls('AppLoading', "material-icons")}>share</em>
         } else {
-            if (!sessionId) {
+            if (!user.loggedIn) {
                 return (
                     <Login callback={ loginHandler }></Login>
                 )
             } else {
                 return (
                     <div className="AppContainer">
-                        <Header user={user} callback={ logoutHandler } themeCallback={setTheme}></Header>
-                        <Comander sessionId={sessionId}></Comander>
+                        <Header toggleMenu={toggleMenu}></Header>
+                        <Comander></Comander>
+                        {menu ? <Sidebar className="animated fadeInLeft" /> : null}
                     </div>
                 )
             }
         }
     }
 
-    return <ThemeContext.Provider value={themes[theme]}>
-        <div className={cls('App', theme)}>
+    return <div className={cls('App', theme.id, theme.size)}>
             { render() }
         </div>
-    </ThemeContext.Provider>
 }
 
 export default App;