+++ src/assets/visualization/Graph.jsx
@@ -1,10 +1,11 @@
-import React, { Component } from "react";
+import React, { useEffect, useCallback, useRef, useState } from "react";
 import { createGraph, mapRelationships, getGraphStats } from "./mapper";
 import { GraphEventHandler } from "./GraphEventHandler";
-import { StyledSvgWrapper } from "./styled";
+import { StyledSvgWrapper, StyledZoomHolder, StyledZoomButton, StyledZoomIcon } from "./styled";
 import graphView from "./components/graphView";
 
 import * as d3 from "d3";
+import { cls } from "../../global/utils";
 
 const dim = {
     // Editor bar
@@ -27,94 +28,112 @@ d3.selection.enter.prototype.appendSVG = function (SVGString) {
     });
 };
 
-export class GraphComponent extends Component {
-    state = {
-        zoomInLimitReached: true,
-        zoomOutLimitReached: false,
-    };
+function GraphComponent(props) {
+    const [svgElement, setSvgElement] = useState(null);
+    const [zoomInLimitReached, setZoomInLimitReached] = useState(true);
+    const [zoomOutLimitReached, setZoomOutLimitReached] = useState(false);
+    const _graph = useRef(null);
+    const _graphView = useRef(null);
+    const _graphEH = useRef(null);
 
-    graphInit(el) {
-        this.svgElement = el;
-    }
-
-    zoomInClicked(el) {
-        const limits = this.graphView.zoomIn(el);
-        this.setState({
-            zoomInLimitReached: limits.zoomInLimit,
-            zoomOutLimitReached: limits.zoomOutLimit,
-        });
-    }
-
-    zoomOutClicked(el) {
-        const limits = this.graphView.zoomOut(el);
-        this.setState({
-            zoomInLimitReached: limits.zoomInLimit,
-            zoomOutLimitReached: limits.zoomOutLimit,
-        });
-    }
-
-    getVisualAreaHeight() {
-        return this.props.frameHeight - dim.frameStatusbarHeight || this.svgElement.parentNode.offsetHeight;
-    }
-
-    componentDidMount() {
-        if (this.svgElement != null) {
-            this.initGraphView();
-            this.graph && this.props.setGraph && this.props.setGraph(this.graph);
-            this.props.getAutoCompleteCallback && this.props.getAutoCompleteCallback(this.addInternalRelationships);
-            this.props.assignVisElement && this.props.assignVisElement(this.svgElement, this.graphView);
+    const addInternalRelationships = useCallback((internalRelationships) => {
+        if (_graph.current) {
+            _graph.current.addInternalRelationships(mapRelationships(internalRelationships, _graph.current));
+            props.onGraphModelChange(getGraphStats(_graph.current));
+            _graphView.current.update();
+            _graphEH.current.onItemMouseOut();
         }
-    }
+    }, [props]);
 
-    initGraphView() {
-        if (!this.graphView) {
+    const getVisualAreaHeight = useCallback(() => {
+        return props.frameHeight - dim.frameStatusbarHeight || svgElement.parentNode.offsetHeight;
+    }, [props.frameHeight, svgElement])
+
+    const initGraphView = useCallback(() => {
+        if (!_graphView.current) {
             const NeoConstructor = graphView;
             const measureSize = () => {
                 return {
-                    width: this.svgElement.offsetWidth,
-                    height: this.getVisualAreaHeight(),
+                    width: svgElement.offsetWidth,
+                    height: getVisualAreaHeight(),
                 };
             };
-            this.graph = createGraph(this.props.nodes, this.props.relationships);
-            this.graphView = new NeoConstructor(this.svgElement, measureSize, this.graph, this.props.graphStyle);
-            this.graphEH = new GraphEventHandler(
-                this.graph,
-                this.graphView,
-                this.props.getNodeNeighbours,
-                this.props.onItemMouseOver,
-                this.props.onItemSelect,
-                this.props.onGraphModelChange
+            _graph.current = createGraph(props.nodes, props.relationships);
+            _graphView.current = new NeoConstructor(svgElement, measureSize, _graph.current, props.graphStyle);
+            _graphEH.current = new GraphEventHandler(
+                _graph.current,
+                _graphView.current,
+                props.getNodeNeighbours,
+                props.onItemMouseOver,
+                props.onItemSelect,
+                props.onGraphModelChange
             );
-            this.graphEH.bindEventHandlers();
-            this.props.onGraphModelChange(getGraphStats(this.graph));
-            this.graphView.resize();
-            this.graphView.update();
+            _graphEH.current.bindEventHandlers();
+            props.onGraphModelChange(getGraphStats(_graph.current));
+            _graphView.current.resize();
+            _graphView.current.update();
         }
+        _graph.current && props.setGraph && props.setGraph(_graph.current);
+        props.getAutoCompleteCallback && props.getAutoCompleteCallback(addInternalRelationships);
+        props.assignVisElement && props.assignVisElement(svgElement, _graphView.current);
+    }, [getVisualAreaHeight, addInternalRelationships, props, svgElement]);
+
+    useEffect(() => {
+        if (svgElement) {
+            initGraphView();
+        }
+    }, [svgElement, initGraphView]);
+
+    useEffect(() => {
+        if (_graphView.current) {
+            _graphView.current.update();
+        }
+    }, [props.graphStyle.version]);
+
+    useEffect(() => {
+        if (_graphView.current) {
+            _graphView.current.resize();
+        }
+    }, [props.fullscreen, props.frameHeight]);
+
+    const zoomInClicked = (el) => {
+        const limits = _graphView.current.zoomIn(el)
+        setZoomInLimitReached(limits.zoomInLimit);
+        setZoomOutLimitReached(limits.zoomOutLimit);
     }
 
-    addInternalRelationships = (internalRelationships) => {
-        if (this.graph) {
-            this.graph.addInternalRelationships(mapRelationships(internalRelationships, this.graph));
-            this.props.onGraphModelChange(getGraphStats(this.graph));
-            this.graphView.update();
-            this.graphEH.onItemMouseOut();
-        }
-    };
-
-    componentDidUpdate(prevProps) {
-        if (prevProps.graphStyle.getVersion() !== this.props.graphStyle.getVersion()) {
-            this.graphView.update();
-        }
-        if (this.props.fullscreen !== prevProps.fullscreen || this.props.frameHeight !== prevProps.frameHeight) {
-            this.graphView.resize();
-        }
+    const zoomOutClicked = (el) => {
+        const limits = _graphView.current.zoomOut(el)
+        setZoomInLimitReached(limits.zoomInLimit);
+        setZoomOutLimitReached(limits.zoomOutLimit);
     }
 
-    render() {
-        return (
-            <StyledSvgWrapper>
-                <svg className="neod3viz" ref={this.graphInit.bind(this)} />
-            </StyledSvgWrapper>
-        );
+    const zoomButtons = () => {
+        if (props.fullscreen || props.zoomEnabled) {
+            return (
+                <StyledZoomHolder>
+                    <StyledZoomButton onClick={zoomInClicked}>
+                        <StyledZoomIcon className={cls(zoomInLimitReached ? "disabled" : "", "material-icons")}>
+                            zoom_in
+                        </StyledZoomIcon>
+                    </StyledZoomButton>
+                    <StyledZoomButton onClick={zoomOutClicked}>
+                        <StyledZoomIcon className={cls(zoomOutLimitReached ? "disabled" : "", "material-icons")}>
+                            zoom_out
+                        </StyledZoomIcon>
+                    </StyledZoomButton>
+                </StyledZoomHolder>
+            );
+        }
+        return null
     }
+
+    return (
+        <StyledSvgWrapper>
+            <svg className="neod3viz" ref={el => setSvgElement(el)} />
+             {zoomButtons()}
+        </StyledSvgWrapper>
+    );
 }
+
+export default GraphComponent;