+++ src/assets/visualization/styled.jsx
@@ -349,24 +349,18 @@ export const StyledInspectorFooterStatusMessage = styled.div`
 
 export const StyledZoomHolder = styled.div`
   position: absolute;
-  bottom: 39px;
+  top: 20px;
   right: 0;
-  padding: 6px 6px 0 6px;
-  border-left: #e6e9ef solid 1px;
-  border-top: #e6e9ef solid 1px;
-  background: #fff;
 `
 
 export const StyledZoomButton = styled.button`
   display: list-item;
   list-style-type: none;
   font-size: 2em;
-  margin-bottom: 10px;
   border: none;
   color: #9b9da2;
   background: transparent;
   border-color: black;
-  padding: 2px 6px 3px;
   &:hover {
     color: black;
   }
@@ -381,3 +375,8 @@ export const StyledZoomButton = styled.button`
     }
   }
 `
+
+export const StyledZoomIcon = styled.em`
+  font-size: 18px;
+`
+