+++ src/components/card/Card.js
@@ -1,17 +1,16 @@
-import React, { useState, useCallback, useRef, useEffect, useContext } from 'react';
-import Summary from './Summary';
+import React, { useState, useCallback, useRef, useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import Summary from './components/Summary/Summary';
+import Download from './components/Download/Download';
 
-import { getChart } from '../../service/neo.service';
+import { getQuery } from '../../service/neo.service';
 
 import { ColumnLayout, RowLayout } from '../../global/layouts';
 import Chart from '../../global/components/chart/Chart';
 import { cls } from '../../global/utils';
-import styles from './Card.module.css';
-import neoGraphStyle from '../../global/components/chart/graphStyle';
-import deepmerge from 'deepmerge';
-import { ThemeContext } from '../../global/utils/hooks/theme';
 
-const graphStyle = new neoGraphStyle();
+import styles from './Card.module.css';
+import actions from '../../global/utils/store/actions';
 
 function Card(props) {
     const [results, setResults] = useState(null);
@@ -20,29 +19,34 @@ function Card(props) {
     const [selected, setSelected] = useState(null);
     const [stats, setStats] = useState(null);
     const [expanded, setExpanded] = useState(false);
-    const [fullscreen, setFullscreen] = useState(false);
-    const theme = useContext(ThemeContext)
+    const [download, setDownload] = useState(false);
+    const [theme, fullscreen, user] = useSelector(state => [state.theme, state.theme.fullscreen, state.user]);
+    const graphStyle = useSelector(state => state.graph);
+    const dispatch = useDispatch();
 
-    const graphStyleData = useRef(null);
     const query = useRef('');
+    const visElement = useRef(null);
 
     useEffect(() => {
-        const _graphStyle = graphStyle.toSheet();
-        const rebasedStyles = deepmerge(_graphStyle, { relationship: { 'text-color-internal': theme.relColor }});
-        graphStyle.loadRules(rebasedStyles);
-        graphStyle.update();
-        graphStyleData.current = graphStyle.toSheet();
-    }, [theme.relColor])
+        dispatch(
+            actions.graph.updateStyle({
+                relationship: {
+                    "text-color-external": theme.relColor,
+                    "text-color-internal": theme.relColorInternal,
+                },
+            })
+        );
+    }, [dispatch, theme.relColor, theme.relColorInternal]);
 
     const fecthData = useCallback(async () => {
         try {
-            const results = await getChart(props.sessionId, props.query);
+            const results = await getQuery(user.sessionId, props.query);
             setResults(results);
             setError(null);
         } catch (err) {
             setError(`${ props.query }: ${err}`);
         }
-    }, [props])
+    }, [props.query, user.sessionId])
 
     useEffect(() => {
         if (query.current !== props.query) {
@@ -50,16 +54,19 @@ function Card(props) {
             setResults(null);
             fecthData();
         }
-    }, [props, fecthData]);
+    }, [props.query, fecthData]);
 
     const toggleExpand = () => {
         setExpanded(!expanded);
     }
 
     const toggleFullScreen = () => {
-        props.toggleFullScreen(!fullscreen);
-        setFullscreen(!fullscreen);
+        dispatch(actions.theme.toggleFullScreen())
     };
+
+    const toggleDownload = () => {
+        setDownload(!download);
+    }
 
     const itemHover = (_item) => {
         if (['node', 'relationship'].includes(_item.type)) {
@@ -78,20 +85,22 @@ function Card(props) {
     }
 
     const graphStyleCallback = (style) => {
-        graphStyleData.current = style.toSheet();
+        dispatch(actions.graph.updateStyle(style.toSheet()));
     }
 
     return (
         <ColumnLayout
             className={cls(styles.card, expanded ? styles.expanded : "", fullscreen ? styles.fullscreen : "")}
-            style={{ marginTop: fullscreen ? 85 - props.idx * 500 : 0 }}
         >
             <header className="row middle spaced">
                 <span className={styles.cardTitle}>QUERY</span>
-                <div className={styles.cardQuery} onClick={() => props.restoreQuery(props.query)}>
+                <div className={styles.cardQuery} onClick={() => props.restoreQuery(null, props.query)}>
                     {props.query}
                 </div>
-                <RowLayout dist="middle right">
+                <RowLayout dist="middle right" className={styles.iconContainer}>
+                    <em className="material-icons" title="download" onClick={toggleDownload}>
+                        save_alt
+                    </em>
                     <em
                         className="material-icons"
                         title={fullscreen ? "minimize" : "maximize"}
@@ -100,7 +109,7 @@ function Card(props) {
                         {fullscreen ? "fullscreen_exit" : "fullscreen"}
                     </em>
                     <em
-                        className={cls(styles.expand, "material-icons", fullscreen ? 'disabled' : '')}
+                        className={cls(styles.expand, "material-icons", fullscreen ? "disabled" : "")}
                         title={expanded ? "contract" : "expand"}
                         onClick={fullscreen ? null : toggleExpand}
                     >
@@ -109,25 +118,34 @@ function Card(props) {
                     <em className="material-icons" title="close" onClick={() => props.deleteQuery(props.query)}>
                         close
                     </em>
+                    {download ? (
+                        <Download results={results} vis={visElement.current} className={styles.downloadWrapper} />
+                    ) : null}
                 </RowLayout>
             </header>
             {results ? (
                 <RowLayout className={styles.cardBody}>
-                    <ColumnLayout className={cls(styles.summary, fullscreen || expanded ? styles.summaryWidder : '')}>
-                        <h3>Summary</h3>
-                        <Summary summary={stats} item={item || selected} graphStyle={graphStyle}></Summary>
+                    <ColumnLayout className={cls(styles.summary, fullscreen || expanded ? styles.summaryWidder : "")}>
+                        <ColumnLayout className={cls(styles.summaryContainer, "hideScroll")}>
+                            <h3>Summary</h3>
+                            <Summary summary={stats} item={item || selected}></Summary>
+                        </ColumnLayout>
                     </ColumnLayout>
                     <Chart
                         style={{ width: "100%" }}
                         result={results}
                         maxNeighbours={30}
-                        sessionId={props.sessionId}
                         itemHovered={itemHover}
                         itemSelected={itemSelected}
                         setSummary={setSummary}
-                        graphStyleData={graphStyleData.current}
+                        graphStyle={graphStyle}
                         graphStyleCallback={graphStyleCallback}
                         autoComplete={false}
+                        zoomEnabled={true}
+                        fullscreen={fullscreen}
+                        assignVisElement={(svgElement, graphElement) =>
+                            (visElement.current = { svgElement, graphElement, type: "graph" })
+                        }
                     />
                 </RowLayout>
             ) : (