+++ src/global/utils/hooks/dispatch.js
@@ -1,0 +1,10 @@
+import { useDispatch } from "react-redux"
+
+export const useAsyncDispatch = () => {
+    const dispatch = useDispatch();
+
+    return async (action, method) => {
+        const result = await method;
+        dispatch(action(result));
+    }
+}