+++ src/global/utils/store/actions/db.js
@@ -1,0 +1,12 @@
+import { DBPROPERTIES } from "../reducers/db"
+
+const setProperties = (properties) => {
+    return {
+        type: DBPROPERTIES,
+        properties
+    }
+}
+
+export default {
+    setProperties
+}