+++ src/global/utils/store/actions/user.js
@@ -1,0 +1,20 @@
+import { SETUSER, LOGOUT } from "../reducers/user"
+
+
+const setUser = userObj => {
+    return {
+        type: SETUSER,
+        userData: userObj
+    }
+}
+
+const logOut = () => {
+    return {
+        type: LOGOUT
+    }
+}
+
+export default {
+    setUser,
+    logOut
+}